파라미터 sweep 백테스트 구현 (spec/0001)
- sweep.rs: SweepConfig, SweepResultRow, SweepResult 타입 정의 및 run_sweep() 함수
- simulator.rs: CandleDataCache, fetch_candle_data(), simulate_with_cache() 독립 함수 분리
- config.rs: RawSweepConfig TOML 파싱 + TomlWrapper에 [sweep] 섹션 지원
- output/console.rs: print_sweep_summary() 비교 테이블 (PF, Ret/DD, 이중 ROI)
- output/csv.rs: write_sweep_csv() 메타데이터 헤더 포함
- 예제: zscore_parameter_sweep.rs, strategy.example.toml [sweep] 섹션
- .gitignore: output/ → /output/ (소스 output 모듈 추적 허용)
- 98개 arb-strategy 테스트 통과, clippy 경고 0

+++ crates/arb-strategy/src/zscore/spread.rs
@@ -351,7 +351,7 @@ mod tests {
                 ts,
                 Some(dec(138_000_000, 0)),
                 Some(dec(1380, 0)),
-                Some(dec(100_000 + i as i64 * 10, 0)),
+                Some(dec(100_000 + i * 10, 0)),
             )
             .unwrap();
         }
